Introduced AbstractFolder.
