+++ src/main/java/com/cloudbees/hudson/plugins/folder/FolderProperty.java
@@ -24,49 +24,15 @@
 
 package com.cloudbees.hudson.plugins.folder;
 
-import com.cloudbees.hudson.plugins.folder.health.FolderHealthMetric;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import hudson.ExtensionPoint;
-import hudson.model.AbstractDescribableImpl;
 import hudson.model.Action;
-import hudson.model.Descriptor.FormException;
-import hudson.model.HealthReport;
-import hudson.model.ReconfigurableDescribable;
 import hudson.model.TopLevelItem;
 import hudson.model.TopLevelItemDescriptor;
-import net.sf.json.JSONObject;
-import org.kohsuke.stapler.StaplerOverridable;
-import org.kohsuke.stapler.StaplerRequest;
-
 import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
+import jenkins.model.TransientActionFactory;
 
-public abstract class FolderProperty<C extends Folder> extends AbstractDescribableImpl<FolderProperty<?>>
-        implements ExtensionPoint, ReconfigurableDescribable<FolderProperty<?>> {
-    /**
-     * The {@link Folder} object that owns this property.
-     * This value will be set by the Hudson code.
-     * Derived classes can expect this value to be always set.
-     */
-    protected transient C owner;
-
-    /**
-     * Hook for performing post-initialization action.
-     */
-    protected void setOwner(C owner) {
-        this.owner = owner;
-    }
-
-    /**
-     * Provides stapler override objects to {@link Folder} so that its URL space can be partially
-     * overridden by properties.
-     *
-     * @see StaplerOverridable
-     */
-    public Collection<?> getItemContainerOverrides() {
-        return Collections.emptyList();
-    }
+public abstract class FolderProperty<C extends Folder> extends AbstractFolderProperty<C> {
 
     /**
      * Determines if the parent container is allowed to create a new item of the given type, or copy from
@@ -83,32 +49,6 @@ public abstract class FolderProperty<C extends Folder> extends AbstractDescribab
         return true;
     }
 
-    public FolderProperty reconfigure(StaplerRequest req, JSONObject form) throws FormException {
-        return form == null ? null : getDescriptor().newInstance(req, form);
-    }
-
-    /**
-     * Folder properties can optionally contribute health reports for the folder. These should be reports of the folder
-     * directly, where a report requires iteration of the items in the folder use the {@link #getHealthMetrics()}
-     * in order to prevent multiple iterations of the items in the folder.
-     *
-     * @return the list of health reports.
-     */
-    @NonNull
-    public List<HealthReport> getHealthReports() {
-        return Collections.emptyList();
-    }
-
-    /**
-     * Returns the health metrics contributed by this property.
-     *
-     * @return the health metrics contributed by this property.
-     */
-    @NonNull
-    public List<FolderHealthMetric> getHealthMetrics() {
-        return Collections.emptyList();
-    }
-
     /**
      * {@link hudson.model.Action}s to be displayed in the folder page.
      * <p/>
@@ -118,6 +58,7 @@ public abstract class FolderProperty<C extends Folder> extends AbstractDescribab
      *
      * @return can be empty but never null.
      * @since 3.14
+     * @deprecated Use {@link TransientActionFactory} instead.
      */
     @NonNull
     public Collection<? extends Action> getFolderActions() {
