+++ src/main/java/com/cloudbees/hudson/plugins/folder/health/FolderHealthMetricDescriptor.java
@@ -24,11 +24,10 @@
 
 package com.cloudbees.hudson.plugins.folder.health;
 
-import com.cloudbees.hudson.plugins.folder.Folder;
+import com.cloudbees.hudson.plugins.folder.AbstractFolder;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import hudson.DescriptorExtensionList;
 import hudson.model.Descriptor;
-import hudson.model.Hudson;
 import jenkins.model.Jenkins;
 
 public abstract class FolderHealthMetricDescriptor extends Descriptor<FolderHealthMetric> {
@@ -36,7 +35,8 @@ public abstract class FolderHealthMetricDescriptor extends Descriptor<FolderHeal
      * Returns true if this {@link FolderHealthMetric} type is applicable to the
      * given folder type.
      */
-    public boolean isApplicable(Class<? extends Folder> containerType) {
+    @SuppressWarnings("rawtypes") // erasure
+    public boolean isApplicable(Class<? extends AbstractFolder> containerType) {
         return true;
     }
