+++ src/main/java/com/cloudbees/hudson/plugins/folder/properties/FolderCredentialsProvider.java
@@ -24,9 +24,9 @@
 
 package com.cloudbees.hudson.plugins.folder.properties;
 
-import com.cloudbees.hudson.plugins.folder.Folder;
-import com.cloudbees.hudson.plugins.folder.FolderProperty;
-import com.cloudbees.hudson.plugins.folder.FolderPropertyDescriptor;
+import com.cloudbees.hudson.plugins.folder.AbstractFolder;
+import com.cloudbees.hudson.plugins.folder.AbstractFolderProperty;
+import com.cloudbees.hudson.plugins.folder.AbstractFolderPropertyDescriptor;
 import com.cloudbees.plugins.credentials.Credentials;
 import com.cloudbees.plugins.credentials.CredentialsMatchers;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
@@ -66,6 +66,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import jenkins.model.TransientActionFactory;
 
 /**
  * A store of credentials that can be used as a Stapler opbject.
@@ -78,7 +79,7 @@ public class FolderCredentialsProvider extends CredentialsProvider {
 
     @Override
     public Set<CredentialsScope> getScopes(ModelObject object) {
-        if (object instanceof Folder) {
+        if (object instanceof AbstractFolder) {
             return SCOPES;
         }
         return super.getScopes(object);
@@ -102,8 +103,8 @@ public class FolderCredentialsProvider extends CredentialsProvider {
         List<C> result = new ArrayList<C>();
         if (ACL.SYSTEM.equals(authentication)) {
             while (itemGroup != null) {
-                if (itemGroup instanceof Folder) {
-                    final Folder folder = Folder.class.cast(itemGroup);
+                if (itemGroup instanceof AbstractFolder) {
+                    final AbstractFolder<?> folder = AbstractFolder.class.cast(itemGroup);
                     FolderCredentialsProperty property = folder.getProperties().get(FolderCredentialsProperty.class);
                     if (property != null) {
                         result.addAll(DomainCredentials.getCredentials(
@@ -125,8 +126,8 @@ public class FolderCredentialsProvider extends CredentialsProvider {
 
     @Override
     public CredentialsStore getStore(@CheckForNull ModelObject object) {
-        if (object instanceof Folder) {
-            final Folder folder = Folder.class.cast(object);
+        if (object instanceof AbstractFolder) {
+            final AbstractFolder<?> folder = AbstractFolder.class.cast(object);
             FolderCredentialsProperty property = folder.getProperties().get(FolderCredentialsProperty.class);
             if (property != null) {
                 return property.getStore();
@@ -135,7 +136,7 @@ public class FolderCredentialsProvider extends CredentialsProvider {
         return null;
     }
 
-    public static class FolderCredentialsProperty extends FolderProperty<Folder> {
+    public static class FolderCredentialsProperty extends AbstractFolderProperty<AbstractFolder<?>> {
 
         /**
          * Old store of credentials
@@ -413,20 +414,33 @@ public class FolderCredentialsProvider extends CredentialsProvider {
             return false;
         }
 
-        @NonNull
-        @Override
-        public Collection<? extends Action> getFolderActions() {
-            return Collections.singleton(new CredentialsStoreAction() {
-                @NonNull
-                @Override
-                public CredentialsStore getStore() {
-                    return FolderCredentialsProperty.this.getStore();
+        @SuppressWarnings({"unchecked", "rawtypes"}) // erasure
+        @Extension
+        public static class ActionFactory extends TransientActionFactory<AbstractFolder> {
+            @Override
+            public Class<AbstractFolder> type() {
+                return AbstractFolder.class;
+            }
+            @Override
+            public Collection<? extends Action> createFor(AbstractFolder target) {
+                final FolderCredentialsProperty prop = ((AbstractFolder<?>) target).getProperties().get(FolderCredentialsProperty.class);
+                if (prop != null) {
+                    return Collections.singleton(new CredentialsStoreAction() {
+                        @NonNull
+                        @Override
+                        public CredentialsStore getStore() {
+                            return prop.getStore();
+                        }
+                    });
+                } else {
+                    return Collections.emptySet();
                 }
-            });
+            }
+
         }
 
         @Extension
-        public static class DescriptorImpl extends FolderPropertyDescriptor {
+        public static class DescriptorImpl extends AbstractFolderPropertyDescriptor {
 
             @Override
             public String getDisplayName() {
