+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/help-icon.html
