+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/newView.jelly
