+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/rename.jelly
