+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/sidepanel.jelly
