+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/tasks-new.properties
