+++ src/main/resources/com/cloudbees/hudson/plugins/folder/AbstractFolder/view-index-top.jelly
