+++ src/main/resources/com/cloudbees/hudson/plugins/folder/Folder/_api.jelly
@@ -25,6 +25,6 @@ THE SOFTWARE.
 
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler">
-    <st:include page="/hudson/model/AbstractItem/_api.jelly"/>
+    <st:include page="/com/cloudbees/hudson/plugins/folder/AbstractFolder/_api.jelly"/>
     <st:include page="/hudson/model/ModifiableItemGroup/_api.jelly"/>
 </j:jelly>
