+++ src/main/resources/com/cloudbees/hudson/plugins/folder/Folder/tasks-create.jelly
@@ -1,7 +1,7 @@
 <!--
 The MIT License
 
-Copyright 2013-2015 CloudBees.
+Copyright 2013 CloudBees.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -21,5 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 -->
-<?jelly escape-by-default='true'?>
-<st:include it="${it.primaryView}" page="sidepanel.jelly" xmlns:st="jelly:stapler"/>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:s="/lib/form">
+  <j:set var="url" value="${h.getNearestAncestorUrl(request,it)}"/>
+  <l:task icon="images/24x24/new-package.png" href="${url}/new" title="${%newJob(it.newPronoun)}" permission="${it.CREATE}" />
+</j:jelly>
