+++ src/main/resources/com/cloudbees/hudson/plugins/folder/Folder/tasks-create.properties
@@ -1,0 +1,1 @@
+newJob=New {0}
