--- src/main/resources/com/cloudbees/hudson/plugins/folder/computed/ComputedFolder/configure-entries.jelly
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core"/>
