--- src/main/resources/com/cloudbees/hudson/plugins/folder/computed/ComputedFolder/configure.jelly
@@ -1,53 +1,0 @@
-<!--
-The MIT License
-
-Copyright (c) 2011-2015, CloudBees, Inc.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
--->
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:i="jelly:fmt">
-    <l:layout title="${it.displayName} Config" norefresh="true" permission="${it.EXTENDED_READ}">
-        <st:include page="sidepanel.jelly"/>
-        <!-- TODO <f:breadcrumb-config-outline/> gives ReferenceError: "section" is not defined -->
-        <l:main-panel>
-            <div class="behavior-loading">${%LOADING}</div>
-            <f:form method="post" action="configSubmit" name="config">
-                <j:set var="descriptor" value="${it.descriptor}"/>
-                <j:set var="instance" value="${it}"/>
-                <f:entry title="${%Name}" field="name">
-                    <f:textbox/>
-                </f:entry>
-                <!-- TODO displayNameOrNull as in Folder/configure.jelly -->
-                <f:entry title="${%Description}" help="${app.markupFormatter.helpUrl}">
-                    <f:textarea name="description" value="${it.description}" codemirror-mode="${app.markupFormatter.codeMirrorMode}" codemirror-config="${app.markupFormatter.codeMirrorConfig}" previewEndpoint="/markupFormatter/previewDescription"/>
-                </f:entry>
-                <!-- TODO defaultView, viewsTabBar -->
-                <st:include page="configure-entries.jelly"/>
-                <j:if test="${h.hasPermission(it, it.CONFIGURE)}">
-                    <f:bottomButtonBar>
-                        <f:submit value="${%Save}"/>
-                        <f:apply/>
-                    </f:bottomButtonBar>
-                </j:if>
-            </f:form>
-            <st:adjunct includes="lib.form.confirm"/>
-        </l:main-panel>
-    </l:layout>
-</j:jelly>
