--- src/main/resources/com/cloudbees/hudson/plugins/folder/computed/ComputedFolder/tasks-new.properties
@@ -1,1 +1,0 @@
-delete=Delete {0}
